backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic[email]==2.10.2
pydantic-settings==2.6.1
email-validator==2.1.0
python-multipart==0.0.12
starlette==0.41.2
httpx==0.27.2

# Database / ORM / Migrations
SQLAlchemy==2.0.36
PyMySQL==1.1.1
alembic==1.14.0
mysqlclient==2.2.5
# TiDB Python Client
# tidb-cloud-sdk-python

# Auth / Security
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# LLM Dependencies
openai>=1.12.0
anthropic>=0.18.1
huggingface-hub>=0.20.3
python-dotenv>=1.0.0
tiktoken>=0.6.0

# Dev / QA
pytest==8.3.4
pytest-asyncio==0.23.6
pytest-cov==4.0.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
flake8==7.1.1
requests==2.32.3
structlog==24.4.0
httpx==0.27.2

# Document processing
PyPDF2==3.0.1
pymupdf==1.23.26
